feat(cli): render wrapper action confirmations as cargo-style status lines
Move format_status into baml_term (verb style now gated on stderr, where
Reporter prints, instead of console's stdout default) and use it for the
wrapper's install/select/uninstall/self-update confirmations, which were
plain lowercase sentences on stdout. They now render like baml-cli's
Reporter lines (12-char right-aligned bold purple verb) and move to
stderr, matching cargo's stream conventions. Informational listings
(--version, toolchain status/list) stay plain, like cargo and rustup.
